BackEnd/requirements.txt
fastapi
ollama
qdrant-client
pandas
pydantic
requests
easyocr
uvicorn
python-multipart
python-dotenv
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
ragas
datasets
langchain 
langchain-community
langchain-ollama
openai
openyxl